refactor: trip 更新と expense 作成・更新ロジックをサービス関数に抽出
